+++ test/test.tsx
@@ -80,10 +80,14 @@ import {
   ProgressStep,
   Radio,
   Select,
+  SelectOption as SelectOpt,
+  SelectMenu,
+  SelectGroup,
   SelectToggle,
   Slider,
   Spinner,
   Tabs,
+  TextInput,
   Title,
   Toggle,
   ToggleGroupItem,
@@ -186,9 +190,9 @@ const alertVariantOption = AlertVariant.default;
     onReadFinished={(fileHandle) => readFinishedHandler(fileHandle)}
     onReadStarted={(fileHandle) => readStartedHandler(fileHandle)}
   />
-  <FileUploadField onTextChange={bar => textHandler(bar)} />
+  <FileUploadField onTextChange={(bar) => textHandler(bar)} />
   <FormSelect onChange={(foo, event) => handler(foo, event)} />
-  <FrogIcon size="sm" color="green" noVerticalAlign />
+  <FrogIcon size='sm' color='green' noVerticalAlign />
   <KebabToggle onToggle={} />
   <Label />
   <Label isTruncated />
@@ -201,6 +205,7 @@ const alertVariantOption = AlertVariant.default;
   <MultipleFileUpload onFileDrop={(foo) => handler(foo)} />
   <Nav flyout={"menu"} />
   <Nav variant='horizontal-subnav' />
+  <Nav onSelect={(foo) => handler(foo)} onToggle={(foo) => handler(foo)} />
   <NotificationBadge isRead />
   <NotificationBadge isRead={false} />
   <NotificationBadge isRead={isRead} />
@@ -208,9 +213,9 @@ const alertVariantOption = AlertVariant.default;
   <OptionsMenu></OptionsMenu>
   <NumberInput allowEmptyInput />
   <OverflowMenuDropdownItem index={0} />
-  <Page onPageResize={({obj}) => handler({obj})} />
-  <PageGroup aria-label="tester" />
-  <PageNavigation aria-label="tester" />
+  <Page onPageResize={({ obj }) => handler({ obj })} />
+  <PageGroup aria-label='tester' />
+  <PageNavigation aria-label='tester' />
   <PageSidebar isNavOpen />
   <PageSidebar nav='Content' />
   <PageToggleButton isNavOpen onNavToggle />
@@ -247,6 +252,7 @@ const alertVariantOption = AlertVariant.default;
   <TableComposable />
   <TableHeader />
   <Td select={tdSelectTypeObj} actions={{ disable: false }} />
+  <TextInput onChange={(foo, event) => handler(foo, event)} />
   <Toggle isPrimary onToggle={} />
   <ToggleGroupItem onChange={(foo, event) => handler(foo, event)} />
   <Tooltip reference />